pyproject.toml
[project]
name = "video-index"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.23.0",
    "streamlit>=1.51.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "ruff>=0.14.6",
]
